+++ packages/core/freeform/sandbox/pages/FreeFormDemo.vue
@@ -1,17 +1,17 @@
 <template>
-  <SandboxPage title="Free Form (Mocking Schema)">
+  <SandboxLayout title="Free Form (Mocking Schema)">
     <Form
       class="form"
       :schema="buildMockingSchema()"
       @change="console.log"
     />
-  </SandboxPage>
+  </SandboxLayout>
 </template>
 
-<script lang="ts" setup>
-import SandboxPage from '../SandboxPage.vue'
-import { buildMockingSchema } from '../../fixtures/schemas/free-form-mocking'
-import { Form } from '@kong-ui-public/freeform'
+<script setup lang="ts">
+import { SandboxLayout } from '@kong-ui-public/sandbox-layout'
+import { Form } from '../../src'
+import { buildMockingSchema } from '../../src/components/free-form-mocking'
 </script>
 
 <style scoped>
