+++ packages/core/freeform/src/components/StringField.vue
@@ -89,7 +89,7 @@
 import { AUTOFILL_SLOT, type AutofillSlot } from '@kong-ui-public/forms'
 import { computed, inject, toRef, useAttrs } from 'vue'
 import type { InputProps, LabelAttributes } from '@kong/kongponents'
-import useI18n from '../composables/useFreeformI18n'
+import useI18n from '../composables/useI18n.ts'
 import EnhancedInput from './EnhancedInput.vue'
 import { USE_SECRET_INPUT_KEY } from '../constants'
