--- packages/entities/entities-plugins/fixtures/schemas/free-form-mocking.ts
@@ -1,614 +1,0 @@
-import type {
-  BooleanFieldSchema,
-  ExpressionFieldSchema,
-  FormSchema,
-  NumberLikeFieldSchema,
-  RecordFieldSchema,
-  StringFieldSchema,
-  MapFieldSchema,
-  ArrayLikeFieldSchema,
-  ForeignFieldSchema,
-} from '@kong-ui-public/freeform'
-
-export function buildStringFieldCases(): Array<{ [name: string]: StringFieldSchema }> {
-  return [
-    {
-      string_simple: {
-        type: 'string',
-      },
-    },
-    {
-      string_complex: {
-        type: 'string',
-        required: true,
-        description: 'A complex string field',
-        default: 'default value',
-      },
-    },
-    {
-      string_referenceable: {
-        type: 'string',
-        encrypted: true,
-        referenceable: true,
-      },
-    },
-  ]
-}
-
-function buildNumberFieldCases(): Array<{ [name: string]: NumberLikeFieldSchema }> {
-  return [
-    {
-      number_simple: {
-        type: 'number',
-      },
-    },
-    {
-      number_complex: {
-        type: 'number',
-        required: true,
-        description: 'A complex number field',
-        default: 42,
-        gt: 0,
-      },
-    },
-    {
-      number_between: {
-        type: 'number',
-        between: [1, 100],
-      },
-    },
-    {
-      number_enum: {
-        type: 'integer',
-        one_of: [-1, 0, 1],
-      },
-    },
-    {
-      number_referenceable: {
-        type: 'number',
-        referenceable: true,
-      },
-    },
-  ]
-}
-
-export function buildBooleanFieldCases(): Array<{ [name: string]: BooleanFieldSchema }> {
-  return [
-    {
-      boolean_simple: {
-        type: 'boolean',
-      },
-    },
-    {
-      boolean_complex: {
-        type: 'boolean',
-        required: true,
-        description: 'A complex boolean field',
-        default: true,
-      },
-    },
-  ]
-}
-
-export function buildRecordFieldCases(): Array<{ [name: string]: RecordFieldSchema }> {
-  return [
-    {
-      record_simple: {
-        type: 'record',
-        description: 'A simple record field',
-        fields: [
-          {
-            string_simple: {
-              type: 'string',
-            },
-          },
-        ],
-      },
-    },
-    {
-      record_required: {
-        type: 'record',
-        required: true,
-        description: 'A required record field',
-        fields: [
-          {
-            string_simple: {
-              type: 'string',
-            },
-          },
-        ],
-      },
-    },
-    {
-      record_recursive: {
-        type: 'record',
-        description: 'A recursive record field',
-        fields: [
-          {
-            nested_record: {
-              type: 'record',
-              fields: [
-                {
-                  number_simple: {
-                    type: 'number',
-                  },
-                },
-              ],
-            },
-          },
-        ],
-      },
-    },
-    {
-      record_default: {
-        type: 'record',
-        description: 'A record field with default values',
-        default: {
-          string_simple: 'default value',
-        },
-        fields: [
-          {
-            string_simple: {
-              type: 'string',
-            },
-          },
-        ],
-      },
-    },
-  ]
-}
-
-export function buildRecordWithEntityChecksCase(): Array<{ [name: string]: RecordFieldSchema }> {
-  return [
-    {
-      record_empty_entity_checks: {
-        type: 'record',
-        description: 'A record field with empty entity checks',
-        entity_checks: [],
-        fields: [
-          {
-            field_one: {
-              type: 'string',
-            },
-          },
-          {
-            field_two: {
-              type: 'number',
-            },
-          },
-        ],
-      },
-    },
-    {
-      record_single_entity_check: {
-        type: 'record',
-        description: 'A record field with entity checks',
-        entity_checks: [
-          {
-            at_least_one_of: ['field_one', 'field_two'],
-          },
-        ],
-        fields: [
-          {
-            field_one: {
-              type: 'string',
-            },
-          },
-          {
-            field_two: {
-              type: 'number',
-            },
-          },
-        ],
-      },
-    },
-    {
-      record_multiple_entity_checks: {
-        type: 'record',
-        description: 'A record field with multiple entity checks',
-        entity_checks: [
-          {
-            mutually_required: ['field_one', 'field_two'],
-          },
-          {
-            mutually_exclusive: ['field_three', 'field_four'],
-          },
-        ],
-        fields: [
-          {
-            field_one: {
-              type: 'string',
-            },
-          },
-          {
-            field_two: {
-              type: 'number',
-            },
-          },
-          {
-            field_three: {
-              type: 'boolean',
-            },
-          },
-          {
-            field_four: {
-              type: 'string',
-            },
-          },
-        ],
-      },
-    },
-  ]
-}
-
-export function buildMapFieldCases(): Array<{ [name: string]: MapFieldSchema }> {
-  return [
-    {
-      map_simple: {
-        type: 'map',
-        keys: {
-          type: 'string',
-        },
-        values: {
-          type: 'number',
-        },
-      },
-    },
-    {
-      map_simple_default: {
-        type: 'map',
-        required: true,
-        description: 'A simple map field with default value',
-        keys: {
-          type: 'string',
-        },
-        values: {
-          type: 'string',
-        },
-        default: { foo: 'bar' },
-      },
-    },
-    {
-      map_referenceable: {
-        type: 'map',
-        keys: {
-          type: 'string',
-        },
-        values: {
-          type: 'string',
-          referenceable: true,
-        },
-      },
-    },
-    {
-      map_value_enum: {
-        type: 'map',
-        keys: {
-          type: 'string',
-        },
-        values: {
-          type: 'string',
-          one_of: ['option1', 'option2', 'option3'],
-        },
-      },
-    },
-    {
-      map_value_record: {
-        type: 'map',
-        keys: {
-          type: 'string',
-        },
-        values: {
-          type: 'record',
-          fields: [
-            {
-              nested_field: {
-                type: 'string',
-              },
-            },
-          ],
-        },
-      },
-    },
-    {
-      map_value_array: {
-        type: 'map',
-        keys: {
-          type: 'string',
-        },
-        values: {
-          type: 'array',
-          elements: {
-            type: 'string',
-          },
-        },
-      },
-    },
-    {
-      map_value_map: {
-        type: 'map',
-        keys: {
-          type: 'string',
-        },
-        values: {
-          type: 'map',
-          keys: {
-            type: 'string',
-          },
-          values: {
-            type: 'string',
-          },
-        },
-      },
-    },
-  ]
-}
-
-export function buildEnumFieldCases(): Array<{ [name: string]: StringFieldSchema | NumberLikeFieldSchema | BooleanFieldSchema }> {
-  return [
-    {
-      enum_simple: {
-        type: 'string',
-        one_of: ['option1', 'option2', 'option3'],
-      },
-    },
-    {
-      enum_complex: {
-        type: 'string',
-        required: true,
-        description: 'A complex enum field',
-        one_of: ['option1', 'option2', 'option3'],
-      },
-    },
-    {
-      enum_default: {
-        type: 'string',
-        one_of: ['option1', 'option2', 'option3'],
-        default: 'option2',
-      },
-    },
-    // fixme: this has a bug in number options
-    // {
-    //   enum_number: {
-    //     type: 'number',
-    //     one_of: [1, 2, 3],
-    //   },
-    // },
-    // fixme: this has a bug in boolean options
-    // {
-    //   enum_boolean: {
-    //     type: 'boolean',
-    //     one_of: [true, false],
-    //   },
-    // },
-  ]
-}
-
-export function buildTagFieldCases(): Array<{ [name: string]: ArrayLikeFieldSchema }> {
-  return [
-    {
-      tag_simple: {
-        type: 'set',
-        elements: {
-          type: 'string',
-        },
-      },
-    },
-    {
-      tag_complex: {
-        type: 'set',
-        required: true,
-        description: 'A complex tag field',
-        elements: {
-          type: 'string',
-        },
-        default: ['tag1', 'tag2'],
-      },
-    },
-  ]
-}
-
-export function buildArrayFieldCases(): Array<{ [name: string]: ArrayLikeFieldSchema }> {
-  return [
-    {
-      array_simple: {
-        type: 'array',
-        elements: {
-          type: 'string',
-        },
-      },
-    },
-    {
-      array_complex: {
-        type: 'array',
-        required: true,
-        description: 'A complex array field',
-        elements: {
-          type: 'number',
-        },
-        default: [0],
-      },
-    },
-    // fixme: should support `resetLabelPath` in tests
-    {
-      array_of_records: {
-        type: 'array',
-        elements: {
-          type: 'record',
-          fields: [
-            {
-              foo: {
-                type: 'string',
-              },
-            },
-          ],
-        },
-        description: 'An array of records',
-      },
-    },
-    {
-      array_of_array: {
-        type: 'array',
-        elements: {
-          type: 'array',
-          elements: {
-            type: 'string',
-          },
-        },
-        default: [['foo']],
-      },
-    },
-    {
-      array_of_maps: {
-        type: 'array',
-        elements: {
-          type: 'map',
-          keys: {
-            type: 'string',
-          },
-          values: {
-            type: 'string',
-          },
-        },
-        description: 'An array of maps',
-      },
-    },
-  ]
-}
-
-export function buildForeignFieldCases(): Array<{ [name: string]: ForeignFieldSchema }> {
-  return [
-    {
-      foreign_simple: {
-        type: 'foreign',
-        reference: 'certificates',
-      },
-    },
-    {
-      foreign_complex: {
-        type: 'foreign',
-        reference: 'certificates',
-        description: 'Certificate reference',
-        required: true,
-        default: { id: '09d83195-e882-4b95-bf0f-9f3615fd2a4f' },
-      },
-    },
-  ]
-}
-
-/**
- * Fields the Gateway marks `expressible`: their value may alternatively be
- * supplied as an expression, which takes priority over the plain value. Each
- * one is paired with a twin under the root `expressions` record built by
- * `buildExpressionTwinCases()` — the two halves are only meaningful together.
- *
- * Wrapped in a record because a twin path mirrors the source path with its first
- * segment swapped for `expressions`, so a field at the very root has no twin.
- * Named something other than `config` on purpose: nothing about the convention
- * is specific to the record real plugin schemas happen to use.
- */
-export function buildExpressibleFieldCases(): Array<{ [name: string]: RecordFieldSchema }> {
-  return [
-    {
-      expressible: {
-        type: 'record',
-        required: true,
-        description: 'Fields whose value can also be given as an expression',
-        fields: [
-          {
-            number_expressible: {
-              type: 'number',
-              expressible: true,
-              description: 'A number field with an expression alternative',
-              gt: 0,
-            },
-          },
-          {
-            string_expressible: {
-              type: 'string',
-              expressible: true,
-              description: 'A string field with an expression alternative',
-            },
-          },
-          {
-            array_expressible: {
-              type: 'array',
-              expressible: true,
-              description: 'An array field whose elements each take an expression',
-              elements: {
-                type: 'number',
-              },
-            },
-          },
-          {
-            number_plain: {
-              type: 'number',
-              description: 'Not expressible — renders no expression control',
-            },
-          },
-        ],
-      },
-    },
-  ]
-}
-
-/**
- * The `expressions` record: one twin per expressible field above, at the same
- * relative path. `Form` never renders this record directly — each entry is
- * rendered inline by the field it overrides.
- */
-export function buildExpressionTwinCases(): Array<{ [name: string]: RecordFieldSchema }> {
-  const expression = (
-    kongType: 'number' | 'string',
-  ): ExpressionFieldSchema => ({
-    type: 'string',
-    len_min: 0,
-    len_max: 1024,
-    expressible_kong_type: kongType,
-  })
-
-  return [
-    {
-      expressions: {
-        type: 'record',
-        required: false,
-        fields: [
-          {
-            number_expressible: expression('number', { type: 'number', gt: 0 }),
-          },
-          {
-            string_expressible: expression('string', { type: 'string' }),
-          },
-          {
-            // Mirrors the source array element-wise, so `expressions.array_expressible.0`
-            // overrides `expressible.array_expressible.0`.
-            array_expressible: {
-              type: 'array',
-              elements: expression('number', { type: 'number' }),
-            },
-          },
-        ],
-      },
-    },
-  ]
-}
-
-export function buildMockingSchema(): FormSchema {
-  return {
-    type: 'record',
-    fields: [
-      ...buildStringFieldCases(),
-      ...buildNumberFieldCases(),
-      ...buildBooleanFieldCases(),
-      ...buildRecordFieldCases(),
-      ...buildRecordWithEntityChecksCase(),
-      ...buildMapFieldCases(),
-      ...buildEnumFieldCases(),
-      ...buildTagFieldCases(),
-      ...buildArrayFieldCases(),
-      ...buildForeignFieldCases(),
-      ...buildExpressibleFieldCases(),
-      ...buildExpressionTwinCases(),
-    ],
-  }
-}
